+++ src/main/java/tn/neopolis/ApplicationManager/repositories/UserRepository.java
@@ -8,7 +8,7 @@ import tn.neopolis.ApplicationManager.models.User;
 
 
 @Repository
-public interface UserRepository extends JpaRepository<User,Long> {
+public interface UserRepository extends JpaRepository<User,Integer> {
 
     User findByUserName(String username);
 }
